+++ tests/test_sample.py
@@ -456,7 +456,7 @@ class TestSample:
         decoded = cwt.decode(token, public_key)
         assert 8 in decoded and isinstance(decoded[8], dict)
         assert 1 in decoded[8] and isinstance(decoded[8][1], dict)
-        c = Claims.from_dict(decoded)
+        c = Claims.new(decoded)
         extracted = COSEKey.new(c.cnf)
         try:
             extracted.verify(msg, sig)
@@ -492,7 +492,7 @@ class TestSample:
         decoded = cwt.decode(token, public_key)
         assert 8 in decoded and isinstance(decoded[8], dict)
         assert 2 in decoded[8] and isinstance(decoded[8][2], list)
-        c = Claims.from_dict(decoded)
+        c = Claims.new(decoded)
         extracted = EncryptedCOSEKey.to_cose_key(c.cnf, enc_key)
         assert extracted.kty == 4  # Symmetric
         assert extracted.alg == 5  # HMAC 256/256
@@ -519,7 +519,7 @@ class TestSample:
         decoded = cwt.decode(token, public_key)
         assert 8 in decoded and isinstance(decoded[8], dict)
         assert 3 in decoded[8] and decoded[8][3] == b"pop-key-id-of-cwt-presenter"
-        c = Claims.from_dict(decoded)
+        c = Claims.new(decoded)
         assert c.cnf == "pop-key-id-of-cwt-presenter"
 
     def test_sample_readme_cwt_with_pop_cose_key(self):
@@ -626,7 +626,7 @@ class TestSample:
         assert isinstance(raw[-70003], dict)
         assert raw[-70003]["baz"] == "qux"
         assert raw[-70004] == 123
-        readable = Claims.from_dict(raw)
+        readable = Claims.new(raw)
         assert readable.get(-70001) == "foo"
         assert readable.get(-70002)[0] == "bar"
         assert readable.get(-70003)["baz"] == "qux"
@@ -658,7 +658,7 @@ class TestSample:
             private_key,
         )
         raw = cwt.decode(token, public_key)
-        readable = Claims.from_dict(
+        readable = Claims.new(
             raw,
             private_claim_names={
                 "ext_1": -70001,